Dataset Utilized in Report
This dataset offers a thorough range of socioeconomic and well-being statistics for almost all nations worldwide. Researchers, analysts, and policymakers can examine the correlations between these aspects by utilizing the data on multiple major categories that it contains. The following categories are included in the dataset:

Gross Domestic Product, or GDP, is a measure of a nation's economic production.

Happiness Index Score: An indicator of general well-being and life satisfaction.

Freedom to Make Life Decisions: Data about individual liberty and freedom.

Social Support: Details about support groups and social networks.

Generosity: Data on the generosity of individuals within a country.

Corruption: Indicators of actual or perceived corruption in a nation.